Fix proofs about the shape of CPS programs
Some proofs refer to the third property instead of the first. Fix that.
There's one remaining proof that uses the third property (for the
`ContFun` case).

The first and third properties are easy to confuse but different. The
3rd means "**all** expressions that are in tail position are
nontrivial", and doesn't mean that "**only** the expressions that are in
tail position are nontrivial" — that's the first property.

Given this is confusing enough for us, I've decided to spell out the
equivalent form of these properties, but feel free to take that back out
again.